cukinia-tests: basic_hardware_tests: add test for minimal flash size
This commit adds a test to check that the target has at least
1GB of flash storage. This minimal value can be configured in the
configuration file using `FLASH_MINIMAL_SIZE_GB`.
This test retrieves the root device from the mountpoint and checks
the size of the corresponding partition in `/proc/partitions`.
It also handles the case where the root device is a mapper device
(e.g., LVM or dm-crypt).

SFL: #17871
Change-Id: I41539a47b12c80d1e268887ffa3f5ace32438597
Signed-off-by: Kévin L'hôpital <kevin.lhopital@savoirfairelinux.com>