launch_conformance_tests: add support for serial connection
Add in launch_conformance_tests the call to a script that will
launch the conformance tests on a board connected via serial port.
This script will compress the cukinia and cukinia-tests directories,
copy them to the board, launch the tests.
It will copy back the results if the report should be created or read
the serial line if it should not be created.

SFL: #17868
Change-Id: Ib05a828037f3fc883359ffc09920b5eedac7b3b4
Signed-off-by: Kévin L'hôpital <kevin.lhopital@savoirfairelinux.com>