cukinia-tests: runtime_os_consumption_tests: add test for OS size
Add a test to check that the OS size does not exceed a certain
percentage of the available space on the root device. The root device
major number is retrieved from the `get_root_device_major.sh` script.
The maximum OS size percentage is defined in the
`tests_configuration.conf`

SFL: #17872
Change-Id: Ib58938044299b2972a51866b7041db042b5c5d10
Signed-off-by: Kévin L'hôpital <kevin.lhopital@savoirfairelinux.com>