.github: add GitHub Actions for CI Conformance
The github action is for now checking for shellcheck errors in the
conformance scripts. It will be extended in the future to run the
conformance tests on a target device via SSH.
This commit also adds a badge to the README to show the status of the
GitHub Actions workflow.

SFL: #17869
Change-Id: I5296354a9b94bd4a0ede2d51e2a8e95d1d23112c
Signed-off-by: Kévin L'hôpital <kevin.lhopital@savoirfairelinux.com>