.github: add reusable workflow for ssh testing
The reusable workflow is designed to run conformance tests
on a target device via ssh. It runs conformance tests with report
generation (uploading the report as artifact) and without report
generation.
The uploading of the report is also done if the previous step fails.
This will allow trying to retrieve the report in case the tests
failed.

SFL: #17869
Change-Id: I14f86dda1d885e0762535176eab1385aa9cbe053
Signed-off-by: Kévin L'hôpital <kevin.lhopital@savoirfairelinux.com>