os_requirements_tests: change NTP tests
The previous NTP tests were checking for the presence of
timedatectl and chronyc commands, which may not be available on all
systems. To be sure that NTP is working, we now compare the host UTC
date with the date set on the board. If they match, it means that
NTP is working correctly. This change makes the test more robust and
less dependent on specific tools.
If the host date variable is not set, the test will be skipped.

SFL: #17873
Change-Id: I37cc14b7f10513a2b5a5f413fa95452d9c24238b
Signed-off-by: Kévin L'hôpital <kevin.lhopital@savoirfairelinux.com>