cukinia-tests:kernel-minimization: add test on unneeded kernel framework
To be sure that kernel is minimized and is not using unneeded
framework, we add a test to check that some kernel framework
are not activated. This will help to reduce the attack surface.
For now as the list of unneeded kernel framework is not defined,
we add a test on SOUND framework.

SFL: #17874
Change-Id: Ib614bcfdf5de7c977a7aeebdf2154e0c94422153
Signed-off-by: Kévin L'hôpital <kevin.lhopital@savoirfairelinux.com>