launch_conformance_tests_ssh: add environment variable for ssh and scp
To allow users to pass additional arguments to the `ssh` and `scp`
commands, we introduce two environment variables:
* `CONFORMANCE_SCP_ARGS`: Additional arguments for the `scp` command
* `CONFORMANCE_SSH_ARGS`: Additional arguments for the `ssh` command
This allows for more flexibility in configuring the SSH connection
and file transfer without modifying the script directly.

As the -O for `scp` is the support for the old SCP protocol, it is
removed as it will be less and less used. If needed, users can
add it back using the `CONFORMANCE_SCP_ARGS` environment variable.

SFL: #17866
Change-Id: I416ebac4378192ccf6c83bfd918c6bcac1ef5057
Signed-off-by: Kévin L'hôpital <kevin.lhopital@savoirfairelinux.com>