test-report-pdf: update submodule to latest commit
This updates the `test-report-pdf` submodule to the latest commit.
This is fixing an issue where the compliance matrix failed to be added
if the last xml report to read did have a test id. Now, the compliance
matrix could be added if there is at least one test id in the all the
xml files.

SFL: #17866
Change-Id: I3b2f583caa4735d03948cddfd34031d9c03dc34a
Signed-off-by: Kévin L'hôpital <kevin.lhopital@savoirfairelinux.com>