.github:workflows: modify CI trigger conditions
Trigger the CI workflow on push to the main branch only and on pull
requests (open, synchronized, reopened). pull_request_target trigger is
the only way to access secrets from a pull request from forks. It
will run the workflow with the context of the base branch, which is
necessary to access the secrets. The changes are still tested when the
repository is checkout.

SFL: #17869
Change-Id: Ia31e6b3fcfd410bd7e8fc74ce9e66a90a4f0b534
Signed-off-by: Kévin L'hôpital <kevin.lhopital@savoirfairelinux.com>