GEISA-API-tests: replace mosquitto_loop_start by mosquitto_loop call
As a while loop is already present in the main function to keep
listening to incoming messages, we can just call mosquitto_loop
from there instead of starting a new thread with mosquitto_loop_start.
Mosquitto_loop is called before connecting to the broker to ensure the
connection is properly handled before entering the main loop.

SFL: #19180
Change-Id: I32d61c66f647f9f45e0020bc546fa8305bd671e3
Signed-off-by: Kévin L'hôpital <kevin.lhopital@savoirfairelinux.com>