GEISA-LEE-tests:os_requirements_tests: check getent installation
Getent is a standard tool installed with the glibc package on most Linux
distributions. But it may not be installed on minimal installations.
This commit adds a check to ensure getent is installed before using it
to verify DNS functionality. This avoids having a test fail due to a
missing utility rather than an actual DNS issue.

SFL: #19183
Change-Id: I73cdee68f888449e98c9a1e0455c8f36b8a4bc89
Signed-off-by: Kévin L'hôpital <kevin.lhopital@savoirfairelinux.com>