GEISA-API-tests: change containerization to lxc/overlayfs
GEISA specifications are requiring to run GEISA application in overlay
container and not OCI. Then, change the GEISA API tests containerization
to use lxc with overlayfs instead of podman/docker.

The launching script is now creating a squashfs image of the test
application and transferring it to the board. Keep the podman build
to build the test application in a containerized environment to
facilitate the cross-compilation.

SFL: #21119
Change-Id: If39f31762447356c5f72f18824df6dff37f08249
Signed-off-by: Kévin L'hôpital <kevin.lhopital@savoirfairelinux.com>